Additional cli subcommands

+++ crates/keycrab_core/queries/passwords/delete.sql
@@ -1,1 +1,1 @@
-delete from passwords where rowid = ?
+delete from passwords where rowid = ? and machine_user_id = ?
